--- api/src/db/Updateable.ts
@@ -1,5 +1,0 @@
-import { AttributeValueUpdate } from '@aws-sdk/client-dynamodb';
-
-export default interface Updateable {
-    getUpdated(): Record<string, AttributeValueUpdate>;
-}
